Simplify the inc scanning logic by just using Module::Metadata.
Because ExtUtils::Installed->new->modules will give you the list of
"primary" modules, I don't think it has to check if a submodule is
updated or not. There might be a case it might not be able to track
when a module is split/extracted from other dist, but I think that
will be upgraded as other distributions are upgraded, hopefully.